Threaded drawing module that draws on message
To achieve this, a MPSC channel is used where the drawing thread is the
consumer and all other threads are producers, i.e. everyone sends
drawing commands to the drawing thread. The commands are sent using the
Msg [enum] type, located in drawing.rs.

The keypress drawing callback is also added here. It sends a Msg to the
drawing thread on char, enter and escape.

The main function is also rewritten. A Result combinator is used
throught the lifetime of the user interaction, i.e. the password is
retreived and then passed on to a couple of Result combinators for
validation and drawing. This is a rough sketch of the logic:

```
read pass -> Validate -ok-> draw(ok) && write pass to file && stop
                      |
                     nok
                      |
                      V
                  draw(fail)
```

Some code maintenance:
 - Larger code blocks are moved into functions.
 - Password related functions are moved to the passwd module.