Dummy draw Enter, Char and Escape
Resolves #4

The actual drawing functions need to be implemented overall, but let's
leave it for the near future. This code implements the messages passed
to the drawing thread and empty () -> () functions that will do the
actual drawing later on.

NOTE:
The passwd module needed a type update on the keypress feedback method.